+++ src/test/java/org/opensearch/knn/memoryoptsearch/faiss/RandomEntryPointsKnnSearchStrategyTests.java
@@ -36,4 +36,26 @@ public class RandomEntryPointsKnnSearchStrategyTests {
             assertTrue(internalVectorId < numVectors);
         }
     }
+
+    @Test
+    @SneakyThrows
+    public void testRandomEntryPointsGeneration_whenNumVectorsIsLessThanNumEntries_thenSuccess() {
+        final int numEntries = 100;
+        final long numVectors = 10;
+
+        // Create strategy
+        final FaissMemoryOptimizedSearcher.RandomEntryPointsKnnSearchStrategy strategy =
+            new FaissMemoryOptimizedSearcher.RandomEntryPointsKnnSearchStrategy(numEntries, numVectors, mock(KnnSearchStrategy.class));
+
+        // Validate #entry points
+        assertEquals(numEntries, strategy.numberOfEntryPoints());
+
+        // We should get exactly `numEntries` vector ids.
+        final DocIdSetIterator iterator = strategy.entryPoints();
+        for (int i = 0; i < numVectors; ++i) {
+            final int internalVectorId = iterator.nextDoc();
+            assertTrue(internalVectorId >= 0);
+            assertTrue(internalVectorId < numVectors);
+        }
+    }
 }